+++ package.json
@@ -47,11 +47,11 @@
     "@storybook/addon-a11y": "^6.3.7",
     "@storybook/addon-essentials": "^6.3.7",
     "@storybook/react": "^6.3.7",
-    "@testing-library/dom": "^7.31.0",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
+    "@testing-library/dom": "^8.11.1",
+    "@testing-library/jest-dom": "^5.15.1",
+    "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.2",
-    "@testing-library/user-event": "^13.2.1",
+    "@testing-library/user-event": "^13.5.0",
     "alex": "^8.0.0",
     "babel-jest": "^27.3.1",
     "babel-loader": "^8.2.3",
@@ -90,7 +90,7 @@
     "react-docgen": "^5.0.0",
     "react-router-enzyme-context": "^1.2.0",
     "react-styleguidist": "^11.0.5",
-    "react-test-renderer": "^16.9.0",
+    "react-test-renderer": "^17.0.2",
     "rollup": "^2.10.4",
     "rollup-plugin-auto-external": "^2.0.0",
     "rollup-plugin-babel": "^4.0.0-beta.2",
