+++ packages/wonder-blocks-dropdown/src/components/__tests__/dropdown-core.test.js
@@ -8,6 +8,8 @@ import SearchTextInput from "../search-text-input.js";
 import DropdownCore from "../dropdown-core.js";
 import {keyCodes} from "../../util/constants.js";
 
+import {unmountAll} from "../../../../../utils/testing/enzyme-shim.js";
+
 jest.mock("../dropdown-core-virtualized.js");
 
 const elementAtIndex = (wrapper, index) =>
@@ -87,6 +89,12 @@ describe("DropdownCore", () => {
     });
 
     afterEach(() => {
+        // We have to explicitly unmount before clearing mocks, otherwise jest
+        // timers will throw because we'll try to clear an animation frame that
+        // we set with a setTimeout but are clearing with clearAnimationFrame
+        // because we restored the clearAnimationFrame mock (and we won't
+        // have cleared the timeout we actually set!)
+        unmountAll();
         jest.restoreAllMocks();
     });
