+++ packages/wonder-blocks-timing/src/util/__tests__/interval.test.js
@@ -7,6 +7,10 @@ describe("Interval", () => {
         jest.useFakeTimers();
     });
 
+    afterEach(() => {
+        jest.restoreAllMocks();
+    });
+
     describe("constructor", () => {
         it("creates instance", () => {
             // Arrange
@@ -44,13 +48,14 @@ describe("Interval", () => {
 
         it("sets an interval when schedule policy is SchedulePolicy.Immediately", () => {
             // Arrange
+            const intervalSpy = jest.spyOn(global, "setInterval");
 
             // Act
             // eslint-disable-next-line no-new
             new Interval(() => {}, 1000, SchedulePolicy.Immediately);
 
             // Assert
-            expect(setInterval).toHaveBeenCalledTimes(1);
+            expect(intervalSpy).toHaveBeenCalledTimes(1);
         });
     });
 
@@ -99,13 +104,18 @@ describe("Interval", () => {
     describe("#set", () => {
         it("should call setInterval", () => {
             // Arrange
-            const interval = new Interval(() => {}, 500);
+            const intervalSpy = jest.spyOn(global, "setInterval");
+            const interval = new Interval(
+                () => {},
+                500,
+                SchedulePolicy.OnDemand,
+            );
 
             // Act
             interval.set();
 
             // Assert
-            expect(setInterval).toHaveBeenNthCalledWith(
+            expect(intervalSpy).toHaveBeenNthCalledWith(
                 1,
                 expect.any(Function),
                 500,
@@ -114,12 +124,15 @@ describe("Interval", () => {
 
         it("should invoke setInterval to call the given action", () => {
             // Arrange
+            const intervalSpy = jest.spyOn(global, "setInterval");
             const action = jest.fn();
-            const interval = new Interval(() => action(), 500);
+            const interval = new Interval(
+                () => action(),
+                500,
+                SchedulePolicy.OnDemand,
+            );
             interval.set();
-            // Flow doesn't know we added jest mocks to this
-            // $FlowFixMe[prop-missing]
-            const scheduledAction = setInterval.mock.calls[0][0];
+            const scheduledAction = intervalSpy.mock.calls[0][0];
 
             // Act
             scheduledAction();
@@ -131,12 +144,16 @@ describe("Interval", () => {
         it("should clear the active interval", () => {
             // Arrange
             const action = jest.fn();
-            const interval = new Interval(() => action(), 500);
+            const interval = new Interval(
+                () => action(),
+                500,
+                SchedulePolicy.OnDemand,
+            );
             interval.set();
 
             // Act
             interval.set();
-            jest.runTimersToTime(501);
+            jest.advanceTimersByTime(501);
 
             // Assert
             expect(action).toHaveBeenCalledTimes(1);
@@ -149,7 +166,7 @@ describe("Interval", () => {
             interval.set();
 
             // Act
-            jest.runTimersToTime(1501);
+            jest.advanceTimersByTime(1501);
 
             // Assert
             expect(action).toHaveBeenCalledTimes(3);
@@ -165,7 +182,7 @@ describe("Interval", () => {
 
             // Act
             interval.clear();
-            jest.runTimersToTime(501);
+            jest.advanceTimersByTime(501);
 
             // Assert
             expect(action).not.toHaveBeenCalled();
@@ -179,7 +196,7 @@ describe("Interval", () => {
 
             // Act
             interval.clear(ClearPolicy.Resolve);
-            jest.runTimersToTime(501);
+            jest.advanceTimersByTime(501);
 
             // Assert
             expect(action).toHaveBeenCalledTimes(1);
@@ -196,7 +213,7 @@ describe("Interval", () => {
 
             // Act
             interval.clear(ClearPolicy.Cancel);
-            jest.runTimersToTime(501);
+            jest.advanceTimersByTime(501);
 
             // Assert
             expect(action).not.toHaveBeenCalled();
@@ -209,7 +226,7 @@ describe("Interval", () => {
 
             // Act
             interval.clear(ClearPolicy.Resolve);
-            jest.runTimersToTime(501);
+            jest.advanceTimersByTime(501);
 
             // Assert
             expect(action).not.toHaveBeenCalled();
